requirements-minimal.txt
# Minimal dependencies (no liboqs, pure Python implementations)
cryptography>=41.0.0
pycryptodome>=3.19.0
pytest>=7.4.0
numpy>=1.24.0
